Extract LabelsBuilder from Labels
Having `LabelsBuilder` outside of `Labels` allows to aggregate the
labels separately, and then build them only once.

That allows to get rid of `Result`s being returned from
`store.add_value()` and `Sample::new()` functions.

It also comes with implication that all labels in `Labels` are already
checked and are valid. Because of this, it's not allowed to add labels
directly to a `Labels` object anymore.